Got rid of NewWarrantyStepTwoViewModel and other useless files. From now on, NewWarrantyViewModel is the reference for both NewWarrantyViewController and NewWarrantyStepTwoViewController.
Fixed a problem in WarrantiesCoordinator, in method showNextStepNewWarrantyScreen that made touching "next step" impossible at first try. Had to switch two lines.

Created a notification in order to refresh warrantiesCollectionView upon saving a new warranty (viewWillAppear was not read and warrantiesCollectionView could not be refreshed, had to find another method).

Started to configure image selection in NewWarrantyNextStepViewController (created alert to display camera/library choices). Now need to handle the privacy settings to be able to open library and camera.