Warranty cells had constraints and consumption issues : constraints because main stackview had a bottom constraint that was conflicting with spacing. Regarding consumption, the problem was due to 'self.layer.shadowPath = UIBezierPath(rect: self.bounds).cgPath' not being called. Had to set this shadowPath property so that iOS doesn't need to calculate transparency dynamically.
Further improvements were made on NewWarrantyController so the warranty end date is calculated according to the starting date plus the value that the user defines through week, month and year steppers. Yet, one last minor problem needs to be solved : adding one year works fine but when adding a new month afterwards, the default datePicker value is used again and month only adds up to the default value, not the updated value after adding a year.