Finally sorted out an issue regarding interaction between warranty's start date, warranty's end date, and steppers (which have an impact on warranty's end date). So now everything is working fine, with the ability to choose a date, increase it with steppers, change the initial start date again, re-increase it with steppers, decrease it, etc. All cases tested were fine.
Now that the warranty's end date is properly working, we've been able to calculate the remaining days until warranty expiry date, and show everything in the warranty cell. 

Adding a photo works fine, we've made sure that all images sizes are re-calculated to 250x320 thanks to an UIImage extension "Resize Image".