Deleted one useless model file.
Moved extension to add a time interval to a date : from NewWarrantyViewController to the Extensions folder.

Fixed some class and method names that were odd.

Created EditWarranty controllers and their dedicated viewmodel. Clicking on the "Edit warranty" button in the WarrantyDetails screen opens a modal controller (the same as the NewWarranty modal), but we passed the corresponding warranty so all the info shows when opening the edit screen. Further configuration is needed for step two (photo controller), and saving the modified fields.