Implemented DIFFABLE data sources instead of regular ones for both Categories and Warranties collection views in HomeWarrantiesListVC.
Started working again on Categories. Improved UI, but still not functional. 

Will first and foremost try to move all things relative to Categories in a new, dedicated file "CategoriesVC". Aim is to make HomeWarrantiesListVC shorter. In that sense, an extension of UIViewController containing "start()" and "remove()" was added to be able to add the child CategoriesVC to the view.