Efficiently saving data when ExtraInfo screen is edited by the user
Needed to find a way to save all the EDITED info given by the user in ExtraInfo edition screen. First, that induced localising each cell from the tableView, to be able to take the textField values (price, model, serial number...) out and save them. But because the table view cells are plain text fields, delegate method "didSelectRowAt" was never triggered. Doing so wasn't an option. Only way I found was to rely on the itemIdentifier in configureExtraInfoTableViewDataSource's closure, and add a target for each textfield upon cell configuration. It triggers a dedicated objc method that instantaneously updates the value held in editWarrantyViewModel, and then in persistent container through properties' didSet. This way, we don't create a gigantic "saveEditedWarranty" method that updates all properties. This method would have been massive if we had to declare each tableview cell to be able to extract their values like this :

     @objc func saveWarranty() {
        let cell = tableView.cellForRow(at: IndexPath(row: 0, section: 0)) as? TextFieldTableViewCell
        if let textFieldText = cell?.textField.text {
            if warrantyModalType == .newWarrantyModal {
                newWarrantyViewModel?.price = Double(textFieldText)
                newWarrantyViewModel?.saveWarranty()
            } else {
                editWarrantyViewModel?.price = Double(textFieldText)
                editWarrantyViewModel?.saveEditedWarranty()
            }
        }
    }

As said above, properties are updated instantaneously. This has a second positive aspect : if the user modifies only one entry, not all values are saved again. Just the one he interacted with.