Minor changes to DeletionAlert.
ExtraInfoViewController now handles adding/saving/editing extra warranty info correctly. Will need to refactor because some methods are huge (but it might be the price to pay to have just one file used for both newWarranty and editWarranty screens).

Also, did the changes in ExtraInfoVC and models, so Notes form is handled (needed extra configuration because Notes is a textView, not a textField).