Managed to repair NewWarrantyProductInfo screen's broken logic.
Problem is that a new bug was spotted : clicking twice in a fast manner on the "lifetime warranty" switch triggers a strange behaviour  (years stepper value comes back to 0). 
Adding Configuration method and struct in cell files is maybe the first step to ensure cells' values are re-configured/updated properly every time the switch is pressed.